Convert project to library
