+++ Cargo.toml
@@ -3,7 +3,9 @@ name = "silentpayments"
 version = "0.1.0"
 edition = "2021"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[lib]
+name = "silentpayments"
+crate-type = ["lib"]
 
 [dependencies]
 # bdk = {version = "0.20.0", features= ["all-keys"] }
